chore: implement ci (#21)
* chore: implement ci

* chore: implement ci

* change ci.yml

* resolve requested changes

* write Dockerfile

+++ api/models/filter_boundaries.go
@@ -34,8 +34,8 @@ func NewFilterBoundaries(in *core.LotFilterBoundaries) *models.LotFilterBoundari
 			To:   swag.Int64(int64(in.MonthlyIncomeMax)),
 		},
 		PaybackPeriod: &models.FloatBoundaries{
-			From: swag.Float64(float64(in.PaybackPeriodMin)),
-			To:   swag.Float64(float64(in.PaybackPeriodMax)),
+			From: swag.Float64(in.PaybackPeriodMin),
+			To:   swag.Float64(in.PaybackPeriodMax),
 		},
 	}
 }
